+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changes
+## [1.8] 2019-06-04
+- Stack now contains path strings rather than PathInfo objects
+- Fix bug where literal paths containing square brackets not pushed and popped correctly
+- Paths completions always use `-Force`
 
 ## [1.7] 2019-04-27
 - Support `cd +` and `cd -` as in PS 6.2